YJDH-424: list commit
