+++ frontend/.eslintrc.nextjs.js
@@ -1,48 +1,4 @@
 module.exports = {
   root: true,
-  extends: ['.eslintrc.base.js', 'plugin:@next/next/recommended'],
-  rules: {
-    'jsx-a11y/anchor-is-valid': [
-      'error',
-      {
-        components: ['Link'],
-        specialLink: ['hrefLeft', 'hrefRight'],
-        aspects: ['invalidHref', 'preferButton'],
-      },
-    ],
-    'no-void': 'off',
-    'react/jsx-pascal-case': ['error', { ignore: ['$*'] }],
-    'react/function-component-definition': 'off',
-  },
-  overrides: [
-    {
-      files: ['*.ts'],
-      rules: {
-        'import/prefer-default-export': 'off',
-      },
-    },
-    {
-      files: ['**/__tests__/**', '*.testcafe.ts'],
-      rules: {
-        'testing-library/render-result-naming-convention': 'off',
-        'jest/expect-expect': 'off',
-        'jest/no-done-callback': 'off',
-        'security/detect-non-literal-regexp': 'off',
-        'no-secrets/no-secrets': 'off',
-        'no-await-in-loop': 'off',
-        'no-restricted-syntax': 0,
-      },
-    },
-    {
-      files: ['*.components.ts'],
-      rules: {
-        'security/detect-non-literal-fs-filename': 'off',
-        '@typescript-eslint/explicit-module-boundary-types': 'off',
-        '@typescript-eslint/explicit-function-return-type': 'off',
-        'testing-library/await-async-query': 'off',
-        'security/detect-non-literal-regexp': 'off',
-        'no-secrets/no-secrets': 'off',
-      },
-    },
-  ],
+  extends: ['.eslintrc.base.js', 'plugin:@next/next/recommended']
 };
