+++ frontend/benefit/applicant/src/components/applications/forms/application/step2/utils/validation.ts
@@ -8,9 +8,8 @@ import {
   MAX_PHONE_NUMBER_LENGTH,
   MAX_SHORT_STRING_LENGTH,
   MIN_PHONE_NUMBER_LENGTH,
-  PAY_SUBSIDY_OPTIONS,
 } from 'benefit/applicant/constants';
-import { VALIDATION_MESSAGE_KEYS } from 'benefit-shared/constants';
+import { PAY_SUBSIDY_OPTIONS, VALIDATION_MESSAGE_KEYS } from 'benefit-shared/constants';
 import startOfYear from 'date-fns/startOfYear';
 import { FinnishSSN } from 'finnish-ssn';
 import { TFunction } from 'next-i18next';
@@ -42,7 +41,7 @@ export const getValidationSchema = (t: TFunction) =>
         is: true,
         then: Yup.mixed().oneOf(
           [null, ...PAY_SUBSIDY_OPTIONS],
-          t(VALIDATION_MESSAGE_KEYS.INVALID)
+          t(VALIDATION_MESSAGE_KEYS.INVALID),
         ),
       }),
     [APPLICATION_FIELDS_STEP2_KEYS.APPRENTICESHIP_PROGRAM]: Yup.boolean()
