+++ frontend/benefit/handler/src/utils/test-utils/axios-test-context.ts
@@ -10,7 +10,7 @@ const AxiosTestContext = Axios.create({
 });
 // To use Nock with Axios, you may need to configure Axios to use the Node adapter
 // More info https://github.com/nock/nock#axios
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment,unicorn/prefer-module
 AxiosTestContext.defaults.adapter = require('axios/lib/adapters/http');
 
 export default AxiosTestContext;
