+++ frontend/shared/.eslintignore
@@ -2,4 +2,6 @@ node_modules
 jest.config.js
 .eslintrc.js
 setupTests.ts
+**/next-server.js
+
